Throttle on Unhandled rate limit
Tracked in: #127

- Unhandled rate limit
  - Incident 1
    - Logs: https://gist.github.com/schneems/be8b7ee0e9ec3b1d8d1e83f13e35ce27 
    - link: https://app.circleci.com/pipelines/github/heroku/hatchet/206/workflows/c5b3126a-3b36-4bc5-98fb-77d997432ae2/jobs/597
  - Incident 2
    - Link https://app.circleci.com/pipelines/github/heroku/hatchet/204/workflows/8b77d5f5-0057-484e-989a-dd0dc8d1a88e/jobs/593
  - incident 3
    - link https://app.circleci.com/pipelines/github/heroku/hatchet/216/workflows/bea5f936-8682-4b92-ac75-69abc3a6386c/jobs/632


```
Retrying failed Attempt #1/3 to push for 'hatchet-t-a2291a3e46' due to error: 
Hatchet::App::FailedDeployError Could not deploy 'hatchet-t-a2291a3e46' (default_ruby) using 'Hatchet::GitApp' at path: 'repo_fixtures/repos/default/default_ruby'
if this was expected add `allow_failure: true` to your deploy hash.
Buildpack: nil
Repo: https://git.heroku.com/hatchet-t-a2291a3e46.git
output:
error: RPC failed; HTTP 429 curl 22 The requested URL returned error: 429 Too Many Requests
fatal: the remote end hung up unexpectedly
fatal: the remote end hung up unexpectedly
Everything up-to-date
```

Also renamed the shell throttle spec file to end in `_spec.rb` and decided it's a good place for the app git tests since they're not deploying code and are more unit-test ish and it's good to stub out the sleep in one place.
